Separate ", '', ${, }, tokens
Because of the iterator-design for the tokenizer, it got a lot more
complicated. But at the same time, the rest of the things got less
complicated: For example is interpolation strings now the same as any
other string (or technically, the other way around!)

It's hard to say if it was worth it, but I think it was. Fixes #6