rename value -> literal
That's pretty subjective as there's no one true terminology, but I
think literal is a better term here:

* value usually refers to the result of evaluation, that is, to a
specific pile of bits in the memory

* literal, on the other hand, usually refers to a bit of syntax

Literal is often, but not always, used to describe "atomic"
self-evaluation constructs like ints or string (for example, in the
Rust reference).

Sometimes it is also used for compound things (ie, "collection
literals"), but that's not the meaning we use here.