make BinOp::operator return an optional value
Since rowan produces syntax trees even if there are errors present in
the source, every function that operates on the tree has to assume that
syntax nodes are partially constructed. As a result, every function on
the tree should return an optional value.

This is already the case with most functions, `BinOp::operator` has been
updated to behave the same way.